modularized campaign generation function and started work on DB update middleware
